extract extend to utils
